+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Feature: The port used for the mutating webhook can be configured using the Helm chart setting
   `agentInjector.webhook.port`.
 
-- Feature: A new repeated `--value a.b.c=v` flag was added to the `telepresence helm install` command so that
+- Feature: A new repeated `--set a.b.c=v` flag was added to the `telepresence helm install` command so that
   values can be passed directly from the command line, without first storing them in a file.
 
 - Change: The default port for the mutating webhook is now `443`. It used to be `8443`.